refactor(xxl-conf): 优化配置数据相关接口
- 将 ConfDataController 中的 id 参数类型从 int 改为 long
- 新增 ConfDataLogController 用于管理配置数据日志
- 实现了配置数据日志的页面展示和分页查询功能
- 优化了权限验证逻辑

+++ xxl-conf-admin/src/main/java/com/xxl/conf/admin/controller/biz/ConfDataController.java
@@ -104,7 +104,7 @@ public class ConfDataController {
     @RequestMapping("/load")
     @ResponseBody
     @Permission
-    public Response<ConfData> load(int id){
+    public Response<ConfData> load(long id){
         return confDataService.load(id);
     }
 
@@ -124,7 +124,7 @@ public class ConfDataController {
     @RequestMapping("/delete")
     @ResponseBody
     @Permission
-    public Response<String> delete(@RequestParam("ids[]") List<Integer> ids, HttpServletRequest request){
+    public Response<String> delete(@RequestParam("ids[]") List<Long> ids, HttpServletRequest request){
         return confDataService.delete(ids, loginService.getLoginUser(request), loginService.isAdmin(request));
     }
